Upgrade to Python 3.10
And wxPython 4.2.0, which finally lets us do this.

Remove the pdfViewer hack

Fixed in wxPython 4.2.0, since my PR got merged. Do note that it doesn't
work with py3.10 anyways due to int/float nonsense, but that's already
reported upstream.

.github/workflows/wb_ci.yaml:
ci: Bump all actions

Closes #619

+++ Mopy/bash/balt.py
@@ -2129,8 +2129,8 @@ class ListBoxes(WrappingTextMixin, DialogWindow):
         self.itemMenu.append(_CheckList_SelectAll())
         self.itemMenu.append(_CheckList_SelectAll(False))
         # TODO(inf) de-wx!
-        minWidth = self._native_widget.ToDIP(
-            self._native_widget.GetTextExtent(title)).width * 1.2 + 64
+        minWidth = int(self._native_widget.ToDIP(
+            self._native_widget.GetTextExtent(title)).width * 1.2 + 64)
         self._panel_text.wrap(minWidth) # otherwise expands to max width
         layout = VLayout(border=5, spacing=5, items=[self._panel_text])
         self._ctrls = {}
